requirements.txt
Django==1.6.9
Pillow==2.7.0
South==1.0.2
bootstrap-admin==0.3.3
django-ckeditor==4.4.7
django-selectable==0.9.0
django-smart-selects==1.0.9
django-suit==0.2.12
django-tagging==0.3.4
django-tagging-autocomplete==0.4.2
sorl-thumbnail==12.2
wsgiref==0.1.2
